Combined the two toolbars
