+++ tests/unit/components/cylc/workspace/toolbar.vue.spec.js
@@ -19,7 +19,7 @@ import { createVuetify } from 'vuetify'
 import { createStore } from 'vuex'
 import { mount } from '@vue/test-utils'
 import storeOptions from '@/store/options'
-import Toolbar from '@/components/cylc/workspace/Toolbar.vue'
+import Toolbar from '@/components/cylc/Toolbar.vue'
 import CommandMenuPlugin from '@/components/cylc/commandMenu/plugin'
 import sinon from 'sinon'
 import WorkflowService from '@/services/workflow.service'
